Refactor project initialization process by modularizing prompts and service functions. Update .gitignore to include 'my-app' and 'dist/' directories. Enhance package manager detection logic for improved reliability.
