+++ .gitignore
@@ -46,3 +46,4 @@ yarn-error.log*
 .idea
 
 dist/
+my-app
