+++ utils/getPkgManager.ts
@@ -1,7 +1,6 @@
-export type PackageManager = "npm" | "pnpm" | "yarn" | "bun";
+import type { PackageManager } from "../types";
 
-export const getPkgManager: () => PackageManager = () => {
-    // This environment variable is set by npm and yarn but pnpm seems less consistent
+export const getPkgManager = (): PackageManager => {
     const userAgent = process.env.npm_config_user_agent;
 
     if (userAgent) {
@@ -11,11 +10,8 @@ export const getPkgManager: () => PackageManager = () => {
             return "pnpm";
         } else if (userAgent.startsWith("bun")) {
             return "bun";
-        } else {
-            return "npm";
         }
-    } else {
-        // If no user agent is set, assume npm
-        return "npm";
     }
-};
+    
+    return "npm";
+};
